Use long long instead of long
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@132461 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 81386ed89970e75de622e5dcfef600c98bbe949f

+++ Rec/Rich/RichRecMonitors/src/RichRecTimeMonitor.h
@@ -70,8 +70,8 @@ namespace Rich
       std::vector<ContainedObject*> m_richPIDs;
 
       // Data to compute average processing time per event and PID
-      unsigned long int m_nEvents; ///< Total number of events
-      unsigned long int m_nPIDs;   ///< Total PIDs considered
+      unsigned long long m_nEvents; ///< Total number of events
+      unsigned long long m_nPIDs;   ///< Total PIDs considered
       long double m_totTime;       ///< Log of total processing time
 
       double m_maxTime;         ///< Maximum overall processing time for plots
